refactor R-CNN for autogluon.  (#1166)
* refactor R-CNN for autogluon. MXNet Oct24 works. MXNet master will encounter OOM and index error

* refactor

* refactor to shorten each file

* fix typo

* python2 compatibility

* revert change

* revert

* change rpn channel

* add amp bn guard

* fix bug

* fix mask rcnn

* temp fix for exporting rcnn models

* better fix

* rename var

* update rpn_channels

* fix style

* change wording

* fix collect_params

* small fix for script

* revert collect_params

* style

* rm model from model_store

* change to relative import

* bump version in RCNNs scripts
