+++ gluoncv/model_zoo/rcnn/faster_rcnn/__init__.py
@@ -3,4 +3,5 @@
 from __future__ import absolute_import
 
 from .faster_rcnn import *
+from .predefined_models import *
 from .rcnn_target import RCNNTargetGenerator, RCNNTargetSampler
