+++ scripts/detection/faster_rcnn/train_faster_rcnn.py
@@ -20,7 +20,7 @@ from mxnet import autograd
 from mxnet.contrib import amp
 import gluoncv as gcv
 
-gcv.utils.check_version('0.6.0')
+gcv.utils.check_version('0.7.0')
 from gluoncv import data as gdata
 from gluoncv import utils as gutils
 from gluoncv.model_zoo import get_model
@@ -32,6 +32,7 @@ from gluoncv.utils.metrics.coco_detection import COCODetectionMetric
 from gluoncv.utils.parallel import Parallelizable, Parallel
 from gluoncv.utils.metrics.rcnn import RPNAccMetric, RPNL1LossMetric, RCNNAccMetric, \
     RCNNL1LossMetric
+from gluoncv.data import COCODetection, VOCDetection
 
 try:
     import horovod.mxnet as hvd
@@ -42,6 +43,7 @@ except ImportError:
 def parse_args():
     parser = argparse.ArgumentParser(description='Train Faster-RCNN networks e2e.')
     parser.add_argument('--network', type=str, default='resnet50_v1b',
+                        choices=['resnet18_v1b', 'resnet50_v1b', 'resnet101_v1d'],
                         help="Base network name which serves as feature extraction base.")
     parser.add_argument('--dataset', type=str, default='voc',
                         help='Training dataset. Now support voc and coco.')
@@ -92,11 +94,11 @@ def parse_args():
                         help='Disable mixup training if enabled in the last N epochs.')
 
     # Norm layer options
-    parser.add_argument('--norm-layer', type=str, default=None,
+    parser.add_argument('--norm-layer', type=str, default=None, choices=[None, 'syncbn'],
                         help='Type of normalization layer to use. '
-                             'If set to None, backbone normalization layer will be fixed,'
-                             ' and no normalization layer will be used. '
-                             'Currently supports \'bn\', and None, default is None.'
+                             'If set to None, backbone normalization layer will be frozen,'
+                             ' and no normalization layer will be used in R-CNN. '
+                             'Currently supports \'syncbn\', and None, default is None.'
                              'Note that if horovod is enabled, sync bn will not work correctly.')
 
     # Loss options
@@ -131,6 +133,123 @@ def parse_args():
                         help='KV store options. local, device, nccl, dist_sync, dist_device_sync, '
                              'dist_async are available.')
 
+    # Advanced options. Expert Only!! Currently non-FPN model is not supported!!
+    # Default setting is for MS-COCO.
+    # The following options are only used if --custom-model is enabled
+    subparsers = parser.add_subparsers(dest='custom_model')
+    custom_model_parser = subparsers.add_parser(
+        'custom-model',
+        help='Use custom Faster R-CNN w/ FPN model. This is for expert only!'
+             ' You can modify model internal parameters here. Once enabled, '
+             'custom model options become available.')
+    custom_model_parser.add_argument(
+        '--no-pretrained-base', action='store_true', help='Disable pretrained base network.')
+    custom_model_parser.add_argument(
+        '--num-fpn-filters', type=int, default=256, help='Number of filters in FPN output layers.')
+    custom_model_parser.add_argument(
+        '--num-box-head-conv', type=int, default=4,
+        help='Number of convolution layers to use in box head if '
+             'batch normalization is not frozen.')
+    custom_model_parser.add_argument(
+        '--num-box-head-conv-filters', type=int, default=256,
+        help='Number of filters for convolution layers in box head.'
+             ' Only applicable if batch normalization is not frozen.')
+    custom_model_parser.add_argument(
+        '--num_box_head_dense_filters', type=int, default=1024,
+        help='Number of hidden units for the last fully connected layer in '
+             'box head.')
+    custom_model_parser.add_argument(
+        '--image-short', type=str, default='800',
+        help='Short side of the image. Pass a tuple to enable random scale augmentation.')
+    custom_model_parser.add_argument(
+        '--image-max-size', type=int, default=1333,
+        help='Max size of the longer side of the image.')
+    custom_model_parser.add_argument(
+        '--nms-thresh', type=float, default=0.5,
+        help='Non-maximum suppression threshold for R-CNN. '
+             'You can specify < 0 or > 1 to disable NMS.')
+    custom_model_parser.add_argument(
+        '--nms-topk', type=int, default=-1,
+        help='Apply NMS to top k detection results in R-CNN. '
+             'Set to -1 to disable so that every Detection result is used in NMS.')
+    custom_model_parser.add_argument(
+        '--post-nms', type=int, default=-1,
+        help='Only return top `post_nms` detection results, the rest is discarded.'
+             ' Set to -1 to return all detections.')
+    custom_model_parser.add_argument(
+        '--roi-mode', type=str, default='align', choices=['align', 'pool'],
+        help='ROI pooling mode. Currently support \'pool\' and \'align\'.')
+    custom_model_parser.add_argument(
+        '--roi-size', type=str, default='7,7',
+        help='The output spatial size of ROI layer. eg. ROIAlign, ROIPooling')
+    custom_model_parser.add_argument(
+        '--strides', type=str, default='4,8,16,32,64',
+        help='Feature map stride with respect to original image. '
+             'This is usually the ratio between original image size and '
+             'feature map size. Since the custom model uses FPN, it is a list of ints')
+    custom_model_parser.add_argument(
+        '--clip', type=float, default=4.14,
+        help='Clip bounding box transformation predictions '
+             'to prevent exponentiation from overflowing')
+    custom_model_parser.add_argument(
+        '--rpn-channel', type=int, default=256,
+        help='Number of channels used in RPN convolution layers.')
+    custom_model_parser.add_argument(
+        '--anchor-base-size', type=int, default=16,
+        help='The width(and height) of reference anchor box.')
+    custom_model_parser.add_argument(
+        '--anchor-aspect-ratio', type=str, default='0.5,1,2',
+        help='The aspect ratios of anchor boxes.')
+    custom_model_parser.add_argument(
+        '--anchor-scales', type=str, default='2,4,8,16,32',
+        help='The scales of anchor boxes with respect to base size. '
+             'We use the following form to compute the shapes of anchors: '
+             'anchor_width = base_size * scale * sqrt(1 / ratio)'
+             'anchor_height = base_size * scale * sqrt(ratio)')
+    custom_model_parser.add_argument(
+        '--anchor-alloc-size', type=str, default='384,384',
+        help='Allocate size for the anchor boxes as (H, W). '
+             'We generate enough anchors for large feature map, e.g. 384x384. '
+             'During inference we can have variable input sizes, '
+             'at which time we can crop corresponding anchors from this large '
+             'anchor map so we can skip re-generating anchors for each input. ')
+    custom_model_parser.add_argument(
+        '--rpn-nms-thresh', type=float, default='0.7',
+        help='Non-maximum suppression threshold for RPN.')
+    custom_model_parser.add_argument(
+        '--rpn-train-pre-nms', type=int, default=12000,
+        help='Filter top proposals before NMS in RPN training.')
+    custom_model_parser.add_argument(
+        '--rpn-train-post-nms', type=int, default=2000,
+        help='Return top proposal results after NMS in RPN training. '
+             'Will be set to rpn_train_pre_nms if it is larger than '
+             'rpn_train_pre_nms.')
+    custom_model_parser.add_argument(
+        '--rpn-test-pre-nms', type=int, default=6000,
+        help='Filter top proposals before NMS in RPN testing.')
+    custom_model_parser.add_argument(
+        '--rpn-test-post-nms', type=int, default=1000,
+        help='Return top proposal results after NMS in RPN testing. '
+             'Will be set to rpn_test_pre_nms if it is larger than rpn_test_pre_nms.')
+    custom_model_parser.add_argument(
+        '--rpn-min-size', type=int, default=1,
+        help='Proposals whose size is smaller than ``min_size`` will be discarded.')
+    custom_model_parser.add_argument(
+        '--rcnn-num-samples', type=int, default=512, help='Number of samples for RCNN training.')
+    custom_model_parser.add_argument(
+        '--rcnn-pos-iou-thresh', type=float, default=0.5,
+        help='Proposal whose IOU larger than ``pos_iou_thresh`` is '
+             'regarded as positive samples for R-CNN.')
+    custom_model_parser.add_argument(
+        '--rcnn-pos-ratio', type=float, default=0.25,
+        help='``pos_ratio`` defines how many positive samples '
+             '(``pos_ratio * num_sample``) is to be sampled for R-CNN.')
+    custom_model_parser.add_argument(
+        '--max-num-gt', type=int, default=100,
+        help='Maximum ground-truth number for each example. This is only an upper bound, not'
+             'necessarily very precise. However, using a very big number may impact the '
+             'training speed.')
+
     args = parser.parse_args()
 
     if args.horovod:
@@ -147,9 +266,31 @@ def parse_args():
     elif args.dataset == 'coco':
         args.epochs = int(args.epochs) if args.epochs else 26
         args.lr_decay_epoch = args.lr_decay_epoch if args.lr_decay_epoch else '17,23'
-        args.lr = float(args.lr) if args.lr else 0.01
+        args.lr = float(args.lr) if args.lr else 0.00125
         args.lr_warmup = args.lr_warmup if args.lr_warmup else 1000
         args.wd = float(args.wd) if args.wd else 1e-4
+
+    def str_args2num_args(arguments, args_name, num_type):
+        try:
+            ret = [num_type(x) for x in arguments.split(',')]
+            if len(ret) == 1:
+                return ret[0]
+            return ret
+        except ValueError:
+            raise ValueError('invalid value for', args_name, arguments)
+
+    if args.custom_model:
+        args.image_short = str_args2num_args(args.image_short, '--image-short', int)
+        args.roi_size = str_args2num_args(args.roi_size, '--roi-size', int)
+        args.strides = str_args2num_args(args.strides, '--strides', int)
+        args.anchor_aspect_ratio = str_args2num_args(args.anchor_aspect_ratio,
+                                                     '--anchor-aspect-ratio', float)
+        args.anchor_scales = str_args2num_args(args.anchor_scales, '--anchor-scales', float)
+        args.anchor_alloc_size = str_args2num_args(args.anchor_alloc_size,
+                                                   '--anchor-alloc-size', int)
+    if args.amp and args.norm_layer == 'syncbn':
+        raise NotImplementedError('SyncBatchNorm currently does not support AMP.')
+
     return args
 
 
@@ -285,7 +426,7 @@ class ForwardBackwardTask(Parallelizable):
             gt_label = label[:, :, 4:5]
             gt_box = label[:, :, :4]
             cls_pred, box_pred, roi, samples, matches, rpn_score, rpn_box, anchors, cls_targets, \
-                box_targets, box_masks, _ = net(data, gt_box, gt_label)
+                box_targets, box_masks, _ = self.net(data, gt_box, gt_label)
             # losses of rpn
             rpn_score = rpn_score.squeeze(axis=-1)
             num_rpn_pos = (rpn_cls_targets >= 0).sum()
@@ -322,7 +463,7 @@ class ForwardBackwardTask(Parallelizable):
                 total_loss.backward()
 
         return rpn_loss1_metric, rpn_loss2_metric, rcnn_loss1_metric, rcnn_loss2_metric, \
-               rpn_acc_metric, rpn_l1_loss_metric, rcnn_acc_metric, rcnn_l1_loss_metric
+            rpn_acc_metric, rpn_l1_loss_metric, rcnn_acc_metric, rcnn_l1_loss_metric
 
 
 def train(net, train_data, val_data, eval_metric, batch_size, ctx, args):
@@ -381,7 +522,11 @@ def train(net, train_data, val_data, eval_metric, batch_size, ctx, args):
         os.makedirs(log_dir)
     fh = logging.FileHandler(log_file_path)
     logger.addHandler(fh)
+    if args.custom_model:
+        logger.info('Custom model enabled. Expert Only!! Currently non-FPN model is not supported!!'
+                    ' Default setting is for MS-COCO.')
     logger.info(args)
+
     if args.verbose:
         logger.info('Trainable parameters:')
         logger.info(net.collect_train_params().keys())
@@ -494,13 +639,60 @@ if __name__ == '__main__':
         module_list.append('fpn')
     if args.norm_layer is not None:
         module_list.append(args.norm_layer)
-        if args.norm_layer == 'bn':
+        if args.norm_layer == 'syncbn':
             kwargs['num_devices'] = len(ctx)
+
     num_gpus = hvd.size() if args.horovod else len(ctx)
     net_name = '_'.join(('faster_rcnn', *module_list, args.network, args.dataset))
+    if args.custom_model:
+        args.use_fpn = True
+        net_name = '_'.join(('custom_faster_rcnn_fpn', args.network, args.dataset))
+        if args.norm_layer == 'syncbn':
+            norm_layer = gluon.contrib.nn.SyncBatchNorm
+            norm_kwargs = {'num_devices': len(ctx)}
+            sym_norm_layer = mx.sym.contrib.SyncBatchNorm
+            sym_norm_kwargs = {'ndev': len(ctx)}
+        elif args.norm_layer == 'gn':
+            norm_layer = gluon.nn.GroupNorm
+            norm_kwargs = {'groups': 8}
+            sym_norm_layer = mx.sym.GroupNorm
+            sym_norm_kwargs = {'groups': 8}
+        else:
+            norm_layer = gluon.nn.BatchNorm
+            norm_kwargs = None
+            sym_norm_layer = None
+            sym_norm_kwargs = None
+        if args.dataset == 'coco':
+            classes = COCODetection.CLASSES
+        else:
+            # default to VOC
+            classes = VOCDetection.CLASSES
+        net = get_model('custom_faster_rcnn_fpn', classes=classes, transfer=None,
+                        dataset=args.dataset, pretrained_base=not args.no_pretrained_base,
+                        base_network_name=args.network, norm_layer=norm_layer,
+                        norm_kwargs=norm_kwargs, sym_norm_kwargs=sym_norm_kwargs,
+                        num_fpn_filters=args.num_fpn_filters,
+                        num_box_head_conv=args.num_box_head_conv,
+                        num_box_head_conv_filters=args.num_box_head_conv_filters,
+                        num_box_head_dense_filters=args.num_box_head_dense_filters,
+                        short=args.image_short, max_size=args.image_max_size, min_stage=2,
+                        max_stage=6, nms_thresh=args.nms_thresh, nms_topk=args.nms_topk,
+                        post_nms=args.post_nms, roi_mode=args.roi_mode, roi_size=args.roi_size,
+                        strides=args.strides, clip=args.clip, rpn_channel=args.rpn_channel,
+                        base_size=args.anchor_base_size, scales=args.anchor_scales,
+                        ratios=args.anchor_aspect_ratio, alloc_size=args.anchor_alloc_size,
+                        rpn_nms_thresh=args.rpn_nms_thresh,
+                        rpn_train_pre_nms=args.rpn_train_pre_nms,
+                        rpn_train_post_nms=args.rpn_train_post_nms,
+                        rpn_test_pre_nms=args.rpn_test_pre_nms,
+                        rpn_test_post_nms=args.rpn_test_post_nms, rpn_min_size=args.rpn_min_size,
+                        per_device_batch_size=args.batch_size // num_gpus,
+                        num_sample=args.rcnn_num_samples, pos_iou_thresh=args.rcnn_pos_iou_thresh,
+                        pos_ratio=args.rcnn_pos_ratio, max_num_gt=args.max_num_gt)
+    else:
+        net = get_model(net_name, pretrained_base=True,
+                        per_device_batch_size=args.batch_size // num_gpus, **kwargs)
     args.save_prefix += net_name
-    net = get_model(net_name, pretrained_base=True,
-                    per_device_batch_size=args.batch_size // num_gpus, **kwargs)
     if args.resume.strip():
         net.load_parameters(args.resume.strip())
     else:
@@ -513,7 +705,7 @@ if __name__ == '__main__':
     if args.amp:
         # Cast both weights and gradients to 'float16'
         net.cast('float16')
-        # This layers doesn't support type 'float16'
+        # These layers don't support type 'float16'
         net.collect_params('.*batchnorm.*').setattr('dtype', 'float32')
         net.collect_params('.*normalizedperclassboxcenterencoder.*').setattr('dtype', 'float32')
