+++ scripts/instance/mask_rcnn/train_mask_rcnn.py
@@ -20,7 +20,7 @@ from mxnet import autograd
 from mxnet.contrib import amp
 import gluoncv as gcv
 
-gcv.utils.check_version('0.6.0')
+gcv.utils.check_version('0.7.0')
 from gluoncv import data as gdata
 from gluoncv import utils as gutils
 from gluoncv.model_zoo import get_model
@@ -31,6 +31,7 @@ from gluoncv.utils.metrics.coco_instance import COCOInstanceMetric
 from gluoncv.utils.metrics.rcnn import RPNAccMetric, RPNL1LossMetric, RCNNAccMetric, \
     RCNNL1LossMetric, MaskAccMetric, MaskFGAccMetric
 from gluoncv.utils.parallel import Parallelizable, Parallel
+from gluoncv.data import COCODetection, VOCDetection
 from multiprocessing import Process
 
 try:
@@ -137,6 +138,131 @@ def parse_args():
                         help='KV store options. local, device, nccl, dist_sync, dist_device_sync, '
                              'dist_async are available.')
 
+    # Advanced options. Expert Only!! Currently non-FPN model is not supported!!
+    # Default setting is for MS-COCO.
+    # The following options are only used if custom-model is enabled
+    subparsers = parser.add_subparsers(dest='custom_model')
+    custom_model_parser = subparsers.add_parser(
+        'custom-model',
+        help='Use custom Faster R-CNN w/ FPN model. This is for expert only!'
+             ' You can modify model internal parameters here. Once enabled, '
+             'custom model options become available.')
+    custom_model_parser.add_argument(
+        '--no-pretrained-base', action='store_true', help='Disable pretrained base network.')
+    custom_model_parser.add_argument(
+        '--num-fpn-filters', type=int, default=256, help='Number of filters in FPN output layers.')
+    custom_model_parser.add_argument(
+        '--num-box-head-conv', type=int, default=4,
+        help='Number of convolution layers to use in box head if '
+             'batch normalization is not frozen.')
+    custom_model_parser.add_argument(
+        '--num-box-head-conv-filters', type=int, default=256,
+        help='Number of filters for convolution layers in box head.'
+             ' Only applicable if batch normalization is not frozen.')
+    custom_model_parser.add_argument(
+        '--num_box_head_dense_filters', type=int, default=1024,
+        help='Number of hidden units for the last fully connected layer in '
+             'box head.')
+    custom_model_parser.add_argument(
+        '--image-short', type=str, default='800',
+        help='Short side of the image. Pass a tuple to enable random scale augmentation.')
+    custom_model_parser.add_argument(
+        '--image-max-size', type=int, default=1333,
+        help='Max size of the longer side of the image.')
+    custom_model_parser.add_argument(
+        '--nms-thresh', type=float, default=0.5,
+        help='Non-maximum suppression threshold for R-CNN. '
+             'You can specify < 0 or > 1 to disable NMS.')
+    custom_model_parser.add_argument(
+        '--nms-topk', type=int, default=-1,
+        help='Apply NMS to top k detection results in R-CNN. '
+             'Set to -1 to disable so that every Detection result is used in NMS.')
+    custom_model_parser.add_argument(
+        '--post-nms', type=int, default=-1,
+        help='Only return top `post_nms` detection results, the rest is discarded.'
+             ' Set to -1 to return all detections.')
+    custom_model_parser.add_argument(
+        '--roi-mode', type=str, default='align', choices=['align', 'pool'],
+        help='ROI pooling mode. Currently support \'pool\' and \'align\'.')
+    custom_model_parser.add_argument(
+        '--roi-size', type=str, default='14,14',
+        help='The output spatial size of ROI layer. eg. ROIAlign, ROIPooling')
+    custom_model_parser.add_argument(
+        '--strides', type=str, default='4,8,16,32,64',
+        help='Feature map stride with respect to original image. '
+             'This is usually the ratio between original image size and '
+             'feature map size. Since the custom model uses FPN, it is a list of ints')
+    custom_model_parser.add_argument(
+        '--clip', type=float, default=4.14,
+        help='Clip bounding box transformation predictions '
+             'to prevent exponentiation from overflowing')
+    custom_model_parser.add_argument(
+        '--rpn-channel', type=int, default=256,
+        help='Number of channels used in RPN convolution layers.')
+    custom_model_parser.add_argument(
+        '--anchor-base-size', type=int, default=16,
+        help='The width(and height) of reference anchor box.')
+    custom_model_parser.add_argument(
+        '--anchor-aspect-ratio', type=str, default='0.5,1,2',
+        help='The aspect ratios of anchor boxes.')
+    custom_model_parser.add_argument(
+        '--anchor-scales', type=str, default='2,4,8,16,32',
+        help='The scales of anchor boxes with respect to base size. '
+             'We use the following form to compute the shapes of anchors: '
+             'anchor_width = base_size * scale * sqrt(1 / ratio)'
+             'anchor_height = base_size * scale * sqrt(ratio)')
+    custom_model_parser.add_argument(
+        '--anchor-alloc-size', type=str, default='384,384',
+        help='Allocate size for the anchor boxes as (H, W). '
+             'We generate enough anchors for large feature map, e.g. 384x384. '
+             'During inference we can have variable input sizes, '
+             'at which time we can crop corresponding anchors from this large '
+             'anchor map so we can skip re-generating anchors for each input. ')
+    custom_model_parser.add_argument(
+        '--rpn-nms-thresh', type=float, default='0.7',
+        help='Non-maximum suppression threshold for RPN.')
+    custom_model_parser.add_argument(
+        '--rpn-train-pre-nms', type=int, default=12000,
+        help='Filter top proposals before NMS in RPN training.')
+    custom_model_parser.add_argument(
+        '--rpn-train-post-nms', type=int, default=2000,
+        help='Return top proposal results after NMS in RPN training. '
+             'Will be set to rpn_train_pre_nms if it is larger than '
+             'rpn_train_pre_nms.')
+    custom_model_parser.add_argument(
+        '--rpn-test-pre-nms', type=int, default=6000,
+        help='Filter top proposals before NMS in RPN testing.')
+    custom_model_parser.add_argument(
+        '--rpn-test-post-nms', type=int, default=1000,
+        help='Return top proposal results after NMS in RPN testing. '
+             'Will be set to rpn_test_pre_nms if it is larger than rpn_test_pre_nms.')
+    custom_model_parser.add_argument(
+        '--rpn-min-size', type=int, default=1,
+        help='Proposals whose size is smaller than ``min_size`` will be discarded.')
+    custom_model_parser.add_argument(
+        '--rcnn-num-samples', type=int, default=512, help='Number of samples for RCNN training.')
+    custom_model_parser.add_argument(
+        '--rcnn-pos-iou-thresh', type=float, default=0.5,
+        help='Proposal whose IOU larger than ``pos_iou_thresh`` is '
+             'regarded as positive samples for R-CNN.')
+    custom_model_parser.add_argument(
+        '--rcnn-pos-ratio', type=float, default=0.25,
+        help='``pos_ratio`` defines how many positive samples '
+             '(``pos_ratio * num_sample``) is to be sampled for R-CNN.')
+    custom_model_parser.add_argument(
+        '--max-num-gt', type=int, default=100,
+        help='Maximum ground-truth number for each example. This is only an upper bound, not'
+             'necessarily very precise. However, using a very big number may impact the '
+             'training speed.')
+    custom_model_parser.add_argument(
+        '--target-roi-scale', type=int, default=2,
+        help='Ratio of mask output roi / input roi. '
+             'For model with FPN, this is typically 2.')
+    custom_model_parser.add_argument(
+        '--num-mask-head-convs', type=int, default=4,
+        help='Number of convolution blocks before deconv layer for mask head. '
+             'For FPN network this is typically 4.')
+
     args = parser.parse_args()
     if args.horovod:
         if hvd is None:
@@ -147,6 +273,29 @@ def parse_args():
     args.lr = float(args.lr) if args.lr else (0.00125 * args.batch_size)
     args.lr_warmup = args.lr_warmup if args.lr_warmup else max((8000 / args.batch_size), 1000)
     args.wd = float(args.wd) if args.wd else 1e-4
+
+    def str_args2num_args(arguments, args_name, num_type):
+        try:
+            ret = [num_type(x) for x in arguments.split(',')]
+            if len(ret) == 1:
+                return ret[0]
+            return ret
+        except ValueError:
+            raise ValueError('invalid value for', args_name, arguments)
+
+    if args.custom_model:
+        args.image_short = str_args2num_args(args.image_short, '--image-short', int)
+        args.roi_size = str_args2num_args(args.roi_size, '--roi-size', int)
+        args.strides = str_args2num_args(args.strides, '--strides', int)
+        args.anchor_aspect_ratio = str_args2num_args(args.anchor_aspect_ratio,
+                                                     '--anchor-aspect-ratio', float)
+        args.anchor_scales = str_args2num_args(args.anchor_scales, '--anchor-scales', float)
+        args.anchor_alloc_size = str_args2num_args(args.anchor_alloc_size,
+                                                   '--anchor-alloc-size', int)
+
+    if args.amp and args.norm_layer == 'bn':
+        raise NotImplementedError('SyncBatchNorm currently does not support AMP.')
+
     return args
 
 
@@ -340,7 +489,7 @@ class ForwardBackwardTask(Parallelizable):
             gt_label = label[:, :, 4:5]
             gt_box = label[:, :, :4]
             cls_pred, box_pred, mask_pred, roi, samples, matches, rpn_score, rpn_box, anchors, \
-                cls_targets, box_targets, box_masks, indices = net(data, gt_box, gt_label)
+                cls_targets, box_targets, box_masks, indices = self.net(data, gt_box, gt_label)
             # losses of rpn
             rpn_score = rpn_score.squeeze(axis=-1)
             num_rpn_pos = (rpn_cls_targets >= 0).sum()
@@ -397,8 +546,8 @@ class ForwardBackwardTask(Parallelizable):
                 total_loss.backward()
 
         return rpn_loss1_metric, rpn_loss2_metric, rcnn_loss1_metric, rcnn_loss2_metric, \
-               mask_loss_metric, rpn_acc_metric, rpn_l1_loss_metric, rcnn_acc_metric, \
-               rcnn_l1_loss_metric, rcnn_mask_metric, rcnn_fgmask_metric
+            mask_loss_metric, rpn_acc_metric, rpn_l1_loss_metric, rcnn_acc_metric, \
+            rcnn_l1_loss_metric, rcnn_mask_metric, rcnn_fgmask_metric
 
 
 def train(net, train_data, val_data, eval_metric, batch_size, ctx, logger, args):
@@ -572,9 +721,57 @@ if __name__ == '__main__':
             kwargs['num_devices'] = len(ctx)
     num_gpus = hvd.size() if args.horovod else len(ctx)
     net_name = '_'.join(('mask_rcnn', *module_list, args.network, args.dataset))
+    if args.custom_model:
+        args.use_fpn = True
+        net_name = '_'.join(('mask_rcnn_fpn', args.network, args.dataset))
+        if args.norm_layer == 'bn':
+            norm_layer = gluon.contrib.nn.SyncBatchNorm
+            norm_kwargs = {'num_devices': len(ctx)}
+            sym_norm_layer = mx.sym.contrib.SyncBatchNorm
+            sym_norm_kwargs = {'ndev': len(ctx)}
+        elif args.norm_layer == 'gn':
+            norm_layer = gluon.nn.GroupNorm
+            norm_kwargs = {'groups': 8}
+            sym_norm_layer = mx.sym.GroupNorm
+            sym_norm_kwargs = {'groups': 8}
+        else:
+            norm_layer = gluon.nn.BatchNorm
+            norm_kwargs = None
+            sym_norm_layer = None
+            sym_norm_kwargs = None
+        if args.dataset == 'coco':
+            classes = COCODetection.CLASSES
+        else:
+            # default to VOC
+            classes = VOCDetection.CLASSES
+        net = get_model('custom_mask_rcnn_fpn', classes=classes, transfer=None,
+                        dataset=args.dataset, pretrained_base=not args.no_pretrained_base,
+                        base_network_name=args.network, norm_layer=norm_layer,
+                        norm_kwargs=norm_kwargs, sym_norm_kwargs=sym_norm_kwargs,
+                        num_fpn_filters=args.num_fpn_filters,
+                        num_box_head_conv=args.num_box_head_conv,
+                        num_box_head_conv_filters=args.num_box_head_conv_filters,
+                        num_box_head_dense_filters=args.num_box_head_dense_filters,
+                        short=args.image_short, max_size=args.image_max_size, min_stage=2,
+                        max_stage=6, nms_thresh=args.nms_thresh, nms_topk=args.nms_topk,
+                        post_nms=args.post_nms, roi_mode=args.roi_mode, roi_size=args.roi_size,
+                        strides=args.strides, clip=args.clip, rpn_channel=args.rpn_channel,
+                        base_size=args.anchor_base_size, scales=args.anchor_scales,
+                        ratios=args.anchor_aspect_ratio, alloc_size=args.anchor_alloc_size,
+                        rpn_nms_thresh=args.rpn_nms_thresh,
+                        rpn_train_pre_nms=args.rpn_train_pre_nms,
+                        rpn_train_post_nms=args.rpn_train_post_nms,
+                        rpn_test_pre_nms=args.rpn_test_pre_nms,
+                        rpn_test_post_nms=args.rpn_test_post_nms, rpn_min_size=args.rpn_min_size,
+                        per_device_batch_size=args.batch_size // num_gpus,
+                        num_sample=args.rcnn_num_samples, pos_iou_thresh=args.rcnn_pos_iou_thresh,
+                        pos_ratio=args.rcnn_pos_ratio, max_num_gt=args.max_num_gt,
+                        target_roi_scale=args.target_roi_scale,
+                        num_fcn_convs=args.num_mask_head_convs)
+    else:
+        net = get_model(net_name, pretrained_base=True,
+                        per_device_batch_size=args.batch_size // num_gpus, **kwargs)
     args.save_prefix += net_name
-    net = get_model(net_name, pretrained_base=True,
-                    per_device_batch_size=args.batch_size // num_gpus, **kwargs)
     if args.resume.strip():
         net.load_parameters(args.resume.strip())
     else:
